requirements.txt
# LLM & AI
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20

# UI
streamlit>=1.30.0
gradio>=4.0.0

# Data Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utils
requests>=2.31.0
Jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional (for enhanced features)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4